pyproject.toml
[project]
name = "petstore-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "structlog>=24.2.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "allure-pytest>=2.13.0",
  "faker>=25.0.0",
  "factory-boy>=3.3.0",
  "testcontainers[postgres]>=4.7.0",
  "mypy>=1.10.0",
  "ruff>=0.4.0",
  "black>=24.4.0",
  "pre-commit>=3.7.0",
  "anyio>=4.4.0",
]
perf = [
  "pytest-benchmark>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-p no:allure_pytest"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = [
  "app/db/migrations/*",
  "app/db/session.py",
  "app/models/*",
  "app/repositories/postgres/*",
  "**/__init__.py",
]

[tool.coverage.report]
fail_under = 80

[tool.black]
line-length = 100
target-version = ["py312"]
